Rename Old and New to Big and Small annotation handlers
