+++ src/test/java/com/brunodles/jsoupparser/SmallAnnotation.java
@@ -10,7 +10,7 @@ import org.junit.runners.JUnit4;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(JUnit4.class)
-public class NewWay {
+public class SmallAnnotation {
 
     private final JsoupParser jsoupParser = new JsoupParser(new ResourceUriResolver());
