+++ src/test/java/com/brunodles/jsoupparser/bigannotation/ErrorModel.java
@@ -1,7 +1,8 @@
-package com.brunodles.jsoupparser.doubles;
+package com.brunodles.jsoupparser.bigannotation;
 
 import com.brunodles.jsoupparser.CssSelector;
 import com.brunodles.jsoupparser.collectors.TextElementCollector;
+import com.brunodles.jsoupparser.doubles.*;
 
 public interface ErrorModel {
