+++ src/test/java/com/brunodles/jsoupparser/bigannotation/WhenParseUrl.java
@@ -1,8 +1,7 @@
-package com.brunodles.jsoupparser.JsoupParserTest;
+package com.brunodles.jsoupparser.bigannotation;
 
 import com.brunodles.jsoupparser.JsoupParser;
 import com.brunodles.jsoupparser.UriResolver;
-import com.brunodles.jsoupparser.doubles.SimpleModel;
 import com.brunodles.jsoupparser.doubles.UnknownException;
 import com.brunodles.jsoupparser.exceptions.ResolverException;
 import org.junit.Before;
